Explore solutions for using one Pipeline with different Resources
@pivotal-nader-ziada was looking into adding support for clusters in #177
when he realized that if the cluster is moved out of PipelineParams into
a Resource, this means that to use 1 Pipeline with 2 different clusters,
you'd have to change the Pipeline itself. This actually applies to all
Resources, and goes against the design principle of Decoupling that we
were aiming for.

In this commit we brainstormed solutions to this problem (one of which
is to accept it) and prototyped solutions in yaml in
`examples_pipeline_reuse`.

Co-authored-by: Nader Ziada <nziada@pivotal.io>