Remove docs on running presubmit tests locally
These docs are hard to keep in sync with the changes to the e2e test
scripts invoked by Prow. Afaik no one is trying to run these locally and
removing these docs avoids having to introduce devs to the the
`DOCKER_REPO_OVERRIDE` env var in addition to the `KO_DOCKER_REPO` env
var.

Also trying to make it easier to find testing docs from the
DEVELOPMENT.md but I think it's still probably not obvious :(