Separate task spec resolution from resource resolution
This is so that the PipelineRun reconciler can use the same resource
resolution. When the PipelineRun resolves resources, there are not yet
any TaskRuns to pass along, those get created later.

Seems a bit weird that the resolution function still needs to be passed
the taskspec + name at all really, but we'll see how this goes!

This is part of #213